configured file structure to have all react redux files in client
